implement phpcs and phpstan, upgrade phpunit
closes #16

+++ includes/abstracts/WordPress.php
@@ -11,6 +11,9 @@ defined( 'WPINC' ) || die();
  */
 abstract class WordPress extends Base {
 
+	/**
+	 * @var string
+	 */
 	protected $orientation = null;
 
 	/**
@@ -18,21 +21,23 @@ abstract class WordPress extends Base {
 	 *
 	 * @param string $path
 	 * @param int $count
-	 * @return array
+	 * @return string[]
 	 */
 	protected static function identify_colors( string $path, int $count = 3 ) : array {
 		require_once \Image_Tag\Plugin::inc() . 'class-get-image-most-common-colors.php';
 
-		$util = new \GetImageMostCommonColors;
+		$util    = new \GetImageMostCommonColors;
 		$_colors = $util->Get_Colors( $path, $count );
 
-		if ( empty( $_colors ) )
+		if ( empty( $_colors ) ) {
 			return array();
+		}
 
 		$colors = array();
 
-		foreach ( $_colors as $color => $percentage )
+		foreach ( $_colors as $color => $percentage ) {
 			$colors[] = '#' . $color;
+		}
 
 		return $colors;
 	}
@@ -45,26 +50,44 @@ abstract class WordPress extends Base {
 	 */
 	protected static function generate_lqip( string $path ) : string {
 		$editor = wp_get_image_editor( $path );
-		$size = $editor->get_size();
+
+		if ( is_wp_error( $editor ) ) {
+			return '';
+		}
+
+		$size  = $editor->get_size();
 		$ratio = $size['height'] / $size['width'];
 
 		$resize_width  = 20;
 		$resize_height = 20;
 
-		if ( $size['width'] > $size['height'] )
+		if ( $size['width'] > $size['height'] ) {
 			$resize_height = $resize_width * $ratio;
-		else if ( $size['width'] < $size['height'] )
+		} else if ( $size['width'] < $size['height'] ) {
 			$resize_width = $resize_height * $ratio;
+		}
 
 		$editor->resize( $resize_width, $resize_height );
 		$editor->set_quality( 50 );
 
-		$path = $editor->generate_filename( 'lqip', get_temp_dir() );
+		$path = ( string ) $editor->generate_filename( 'lqip', get_temp_dir() );
+
+		if ( empty( $path ) ) {
+			return '';
+		}
+
 		$editor->save( $path );
 
-		$mime = wp_get_image_mime( $path );
-		$plain_encoded = base64_encode( file_get_contents( $path ) );
-		$data64 = sprintf( 'data:%s;base64,%s', $mime, $plain_encoded );
+		$mime     = wp_get_image_mime( $path );
+		$contents = ( string ) file_get_contents( $path );
+
+		if ( empty( $contents ) ) {
+			return '';
+		}
+
+		$plain_encoded = base64_encode( $contents );
+		$data64        = sprintf( 'data:%s;base64,%s', $mime, $plain_encoded );
+
 		unlink( $path );
 
 		return $data64;
@@ -79,17 +102,15 @@ abstract class WordPress extends Base {
 	protected function set_orientation() : void {
 		$ratio = $this->ratio();
 
-		if ( $ratio > 1 )
+		if ( $ratio > 1 ) {
 			$this->orientation = 'portrait';
-
-		else if ( $ratio < 1 )
+		} else if ( $ratio < 1 ) {
 			$this->orientation = 'landscape';
-
-		else if ( 1 === $ratio )
+		} else if ( floatval( 1 ) === $ratio ) {
 			$this->orientation = 'square';
-
-		else
+		} else {
 			$this->orientation = 'unknown';
+		}
 	}
 
 	/**
@@ -97,16 +118,16 @@ abstract class WordPress extends Base {
 	 *
 	 * @return float
 	 */
-	abstract function ratio() : float;
+	abstract public function ratio() : float;
 
 	/**
 	 * Get most common colors.
 	 *
 	 * @param int $count
 	 * @uses static::identify_colors()
-	 * @return array
+	 * @return string[]
 	 */
-	abstract function colors( int $count = 5 ) : array;
+	abstract public function colors( int $count = 5 ) : array;
 
 	/**
 	 * Get most common color.
@@ -114,7 +135,7 @@ abstract class WordPress extends Base {
 	 * @uses $this->get_colors()
 	 * @return string
 	 */
-	function mode_color() : string {
+	public function mode_color() : string {
 		$colors = $this->colors();
 		return $colors[0];
 	}
@@ -124,6 +145,6 @@ abstract class WordPress extends Base {
 	 *
 	 * @return string
 	 */
-	abstract function lqip() : string;
+	abstract public function lqip() : string;
 
 }
